refactor: split sqlite db logic by domain
